Added very basic memory management with the linear allocator. Supports malloc but not free. WIP

+++ lib/include/stddef.h
@@ -5,6 +5,6 @@
 
 #define NULL 0
 
-typedef uint64_t		size_t;
+typedef uint32_t		size_t;
 
 #endif
